figuring out how to get random value from dict

+++ Day1to10/Day1to10/day10.py
@@ -19,10 +19,10 @@ def perform_operation(a, b):
     else:
         return "Invalid operation selection"
 
-num_first = int(input("What's the first number?: "))
+num_first = float(input("What's the first number?: "))
 print_operations()
 operation_choice = input("Pick an operation: ")
-num_second = int(input("What's the next number?: "))
+num_second = float(input("What's the next number?: "))
 
 continue_calculating = True
 
@@ -37,7 +37,7 @@ while continue_calculating == True:
     num_first = perform_operation(num_first, num_second)
     print_operations()
     operation_choice = input("Pick an operation: ")
-    num_second = int(input("What's the next number?: "))
+    num_second = float(input("What's the next number?: "))
     perform_operation(num_first, num_second)
     print(f"{num_first} {operation_choice} {num_second} = {perform_operation(num_first, num_second)}")
     continue_choice = input(f"Type 'y' to continue calculating with {perform_operation(num_first, num_second)}, or type 'n' to start a new calculation:")
